establishing the paradigm
So, it looks like the variety of types you can expose to JavaScript
are very limited. Since I still never figured out how to call my own
JavaScript from Rust (in contrast to `alert` and `log`), I'm going to
make the JavaScript call on the Rust/WASM for instructions on what to
render, and I've improvised a protocol for that. (First, query for the
number of entities, then query for X/Y/radius/type for each entity.)
Is this at all sane? I don't know what real game developers would do.