work towards the innovation of Patrol AI
I usually consider it beneath my honor to check in commented-out
code, but there's a degree of heartbreak that calls for it.

I know that the smart way to do this will have to look at our current
velocity. There was even a talk at !!Con West just before mine about
how to do this. There's an embarrassment at not already knowing the
simple physics?