up and running!
What better time to start hacking on this again than at RustConf?! I
managed to get the WASM build working on my laptop! Not sure how I
originally got it working on the other machine or the instructions
have changed (the key turned out to be the "file:../pkg" NPM config
line, which you would have expected to already be there given that
I've worked on this before at all).