state-machine Patrol AI behaves mostly reasonably ... to a point
It goes around the loop of waypoints five and half times in a minute
and a half, and then some critical misalignment between "disorient"
and "deaccel" creeps in, and it goes Warp-10 haywire. I must have a
very bad bug somewhere (maybe I'm conflating π and 0 somewhere in a
way that usually doesn't matter, until the figures line up in just the
wrong way??).